queue schema match repo style
