+++ models/mongoose.ts
@@ -8,7 +8,7 @@ import KeyValue from './db/keyValue';
 import Level from './db/level';
 import Notification from './db/notification';
 import PlayAttempt from './db/playAttempt';
-import { QueueMessage } from './db/queueMessage';
+import QueueMessage from './db/queueMessage';
 import Record from './db/record';
 import Review from './db/review';
 import Stat from './db/stat';
@@ -23,7 +23,7 @@ import KeyValueSchema from './schemas/keyValueSchema';
 import LevelSchema from './schemas/levelSchema';
 import NotificationSchema from './schemas/notificationSchema';
 import PlayAttemptSchema from './schemas/playAttemptSchema';
-import { QueueMessageSchema } from './schemas/queueMessageSchema';
+import QueueMessageSchema from './schemas/queueMessageSchema';
 import RecordSchema from './schemas/recordSchema';
 import ReviewSchema from './schemas/reviewSchema';
 import StatSchema from './schemas/statSchema';
