+++ models/schemas/queueMessageSchema.ts
@@ -1,12 +1,20 @@
 import mongoose from 'mongoose';
-import { QueueMessage, QueueMessageState, QueueMessageType } from '../db/queueMessage';
+import QueueMessage from '../db/queueMessage';
 
-export const QueueMessageSchema = new mongoose.Schema<QueueMessage>({
-  priority: {
-    type: Number,
-    required: false,
-    default: 0,
-  },
+export enum QueueMessageState {
+  PENDING = 'pending', // ready to be processed
+  PROCESSING = 'processing',
+  COMPLETED = 'completed',
+  FAILED = 'failed'
+}
+
+export enum QueueMessageType {
+  FETCH = 'FETCH',
+  REFRESH_INDEX_CALCULATIONS = 'REFRESH_INDEX_CALCULATIONS',
+  CALC_PLAY_ATTEMPTS = 'CALC_PLAY_ATTEMPTS',
+}
+
+const QueueMessageSchema = new mongoose.Schema<QueueMessage>({
   dedupeKey: {
     type: String,
     required: false,
@@ -15,44 +23,51 @@ export const QueueMessageSchema = new mongoose.Schema<QueueMessage>({
     type: mongoose.Schema.Types.ObjectId,
     required: false,
   },
+  log: {
+    type: [String],
+    required: false,
+  },
   message: {
     type: String,
     required: true,
   },
-  state: {
-    type: String,
-    required: true,
-    enum: QueueMessageState
-  },
-  type: {
-    type: String,
-    required: true,
-    enum: QueueMessageType
-  },
-  processingStartedAt: {
-    type: Date,
+  priority: {
+    type: Number,
     required: false,
-  },
-  processingCompletedAt: {
-    type: Date,
-    required: false,
+    default: 0,
   },
   processingAttempts: {
     type: Number,
     required: true,
     default: 0,
   },
-  log: {
-    // array of string
-    type: [String],
+  processingCompletedAt: {
+    type: Date,
     required: false,
+  },
+  processingStartedAt: {
+    type: Date,
+    required: false,
+  },
+  state: {
+    type: String,
+    required: true,
+    enum: QueueMessageState,
+  },
+  type: {
+    type: String,
+    required: true,
+    enum: QueueMessageType,
   },
 },
 {
   timestamps: true,
 });
+
 // add indexes
 QueueMessageSchema.index({ state: 1, priority: -1, createdAt: 1 });
 // partial index where state is pending
 QueueMessageSchema.index({ dedupeKey: 1, type: 1 }, { unique: true, partialFilterExpression: { state: QueueMessageState.PENDING } });
 QueueMessageSchema.index({ jobRunId: 1, state: 1, createdAt: 1 });
+
+export default QueueMessageSchema;
